isolate monoid
