+++ src/monoid.rs
@@ -1,0 +1,23 @@
+use cargo_snippet::snippet;
+
+#[snippet("Monoid")]
+#[allow(dead_code)]
+pub trait Monoid {
+    type T: Clone;
+    fn id() -> Self::T;
+    fn op(a: &Self::T, b: &Self::T) -> Self::T;
+}
+
+#[snippet("Monoid-SUM")]
+#[allow(dead_code)]
+pub struct SUM;
+#[snippet("Monoid-SUM")]
+impl Monoid for SUM {
+    type T = u64;
+    fn id() -> Self::T {
+        0
+    }
+    fn op(a: &Self::T, b: &Self::T) -> Self::T {
+        *a + *b
+    }
+}
