+++ src/segtree.rs
@@ -1,13 +1,8 @@
+use crate::monoid::Monoid;
 use cargo_snippet::snippet;
-#[snippet("SEG")]
-#[allow(dead_code)]
-pub trait Monoid {
-    type T: Clone;
-    fn id() -> Self::T;
-    fn op(a: &Self::T, b: &Self::T) -> Self::T;
-}
 
 #[snippet("SEG")]
+#[snippet(include = "Monoid")]
 #[allow(dead_code)]
 /// Segment Tree
 pub struct SEG<M: Monoid> {
@@ -96,22 +91,9 @@ impl<M: Monoid> SEG<M> {
     }
 }
 
-#[snippet("Monoid-SUM")]
-#[allow(dead_code)]
-struct SUM;
-#[snippet("Monoid-SUM")]
-impl Monoid for SUM {
-    type T = u64;
-    fn id() -> Self::T {
-        0
-    }
-    fn op(a: &Self::T, b: &Self::T) -> Self::T {
-        *a + *b
-    }
-}
-
 #[test]
 fn test_segtree_vs_cumulative_sum() {
+    use crate::monoid::SUM;
     use crate::util::random_range;
     use rand::{Rng, SeedableRng, StdRng};
 
@@ -144,6 +126,7 @@ fn test_segtree_vs_cumulative_sum() {
 
 #[test]
 fn test_segtree_same_index() {
+    use crate::monoid::SUM;
     let seg: SEG<SUM> = SEG::new(8);
     assert_eq!(seg.query(0, 0), 0);
 }
@@ -190,6 +173,7 @@ use test::Bencher;
 
 #[bench]
 fn bench_segtree_update(b: &mut Bencher) {
+    use crate::monoid::SUM;
     use rand::{Rng, SeedableRng, StdRng};
 
     let size = 10000;
@@ -218,6 +202,7 @@ fn bench_segtree_update(b: &mut Bencher) {
 
 #[bench]
 fn bench_segtree_query(b: &mut Bencher) {
+    use crate::monoid::SUM;
     use crate::util;
     use rand::{Rng, SeedableRng, StdRng};
